+++ benchmarks/sac_run_experiments.py
@@ -11,23 +11,26 @@ PATH_NOHUP_OUT_DIR.mkdir(parents=True, exist_ok=True)
 
 # 配置任务
 tasks = [
-    # (env_type, env_name, cuda_list, seed_list)
-    ("gymnasium", "Hopper-v4",          [0, 0, 0], [0, 1, 2]),
-    ("gymnasium", "Ant-v4",             [1, 1, 1], [0, 1, 2]),
-    ("gymnasium", "HalfCheetah-v4",     [2, 2, 2], [0, 1, 2]),
-    ("gymnasium", "HumanoidStandup-v4", [3, 3, 3], [0, 1, 2]),
-    ("gymnasium", "Humanoid-v4",        [4, 4, 4], [0, 1, 2]),
-    # # DMC EASY
-    ("dmc", "walker-walk",              [5, 5, 5], [0, 1, 2]),
-    ("dmc", "walker-run",              [5, 5, 5], [0, 1, 2]),
+    # (env_subcommand, env_name, cuda_list, seed_list)
+    ("env:gym", "Ant-v4",                   [1, 1, 1], [0, 1, 2]),
+    ("env:gym", "HalfCheetah-v4",           [2, 2, 2], [0, 1, 2]),
+    ("env:gym", "Hopper-v4",                [0, 0, 0], [0, 1, 2]),
+    ("env:gym", "HumanoidStandup-v4",       [3, 3, 3], [0, 1, 2]),
+    ("env:gym", "Humanoid-v4",              [4, 4, 4], [0, 1, 2]),
+    ("env:gym", "InvertedPendulum-v4",      [0, 0, 0], [0, 1, 2]),
+    ("env:gym", "Pusher-v5",                [0, 0, 0], [0, 1, 2]),
+    ("env:gym", "Walker2d-v4",              [1, 1, 1], [0, 1, 2]),
+    # DMC EASY
+    ("env:dmc", "walker-walk",              [5, 5, 5], [0, 1, 2]),
+    ("env:dmc", "walker-run",               [5, 5, 5], [0, 1, 2]),
     # DMC HARD
-    ("dmc", "humanoid-walk",            [6, 6, 6], [0, 1, 2]),
-    ("dmc", "dog-walk",                 [7, 7, 7], [0, 1, 2]),
-    ("dmc", "humanoid-run",             [1, 1, 1], [0, 1, 2]),
-    ("dmc", "dog-run",                  [1, 1, 1], [0, 1, 2]),
-    ("dmc", "dog-trot",                 [1, 1, 1], [0, 1, 2]),
-    ("dmc", "humanoid-stand",           [0, 0, 0], [0, 1, 2]),
-    ("dmc", "dog-stand",                [1, 1, 1], [0, 1, 2]),
+    ("env:dmc", "humanoid-walk",            [6, 6, 6], [0, 1, 2]),
+    ("env:dmc", "dog-walk",                 [7, 7, 7], [0, 1, 2]),
+    ("env:dmc", "humanoid-run",             [1, 1, 1], [0, 1, 2]),
+    ("env:dmc", "dog-run",                  [1, 1, 1], [0, 1, 2]),
+    ("env:dmc", "dog-trot",                 [1, 1, 1], [0, 1, 2]),
+    ("env:dmc", "humanoid-stand",           [0, 0, 0], [0, 1, 2]),
+    ("env:dmc", "dog-stand",                [1, 1, 1], [0, 1, 2]),
 ]
 
 # 额外参数（可选）
@@ -49,10 +52,10 @@ def run_command(cmd):
 
 if __name__ == "__main__":
     total = 0
-    for env_type, env_name, cuda_list, seed_list in tasks:
+    for env_subcommand, env_name, cuda_list, seed_list in tasks:
         for cuda_id, seed in zip(cuda_list, seed_list):
             cmd = base_cmd + [
-                "--env.env-type", env_type,
+                env_subcommand,
                 "--env.env-name", env_name,
                 "--env.seed", str(seed),
                 "--agent.seed", str(seed),
@@ -60,7 +63,7 @@ if __name__ == "__main__":
             ] + extra_args
 
             # 用 nohup 启动并输出到对应 log 文件
-            log_file = str(PATH_NOHUP_OUT_DIR / f"log_sac_{env_name}_seed{seed}_{time.strftime('%Y%m%d_%H%M%S')}.out")
+            log_file = str(PATH_NOHUP_OUT_DIR / f"sac_{env_subcommand}_{env_name}_seed{seed}_{time.strftime('%Y%m%d_%H%M%S')}.log")
             full_cmd = ["nohup"] + cmd + [">", log_file, "2>&1", "&", "echo", "$!"]
             pid = subprocess.check_output(" ".join(full_cmd), shell=True).decode().strip()
             # os.system(" ".join(full_cmd))