+++ katarl2/envs/__init__.py
@@ -1,1 +1,4 @@
 from katarl2.envs.common.env_cfg import EnvConfig, get_env_name
+from katarl2.envs.env_dmc import DMCEnvConfig
+from katarl2.envs.env_gymnasium import GymAtariEnvConfig, GymMujocoEnvConfig
+from katarl2.envs.env_envpool import EnvpoolAtariConfig